[04/12/2013] M    libcuda/cuda_macro.h 1. [ADD] Include stdlib.h.
M    CMakeLists.txt
1. [MOD] Do not link opencv libraries. Instead, link lodepng.
2. [ADD] Change the order of libdvrsuite.
3. [MOD] Change the name example_dvr2_suite to example_dvr_suite.

M    libdvrsuite/DvrSuiteWin.cpp
1. [MOD] Cast the type of 0 to size_t when calling max() to avoid compiler error.

+++ trunk/libdvrsuite/DvrSuiteWin.cpp
@@ -58,7 +58,9 @@ CDvrSuiteWin::
 	for(size_t v = 0; v < uNrOfElements; v++)
 	{
 		size_t uBin = (size_t)floor((double)(uNrOfBins - 1) * ((double)data[v] - dValueMin) / (double)(dValueMax - dValueMin));
-		uBin = min(max(uBin, 0), uNrOfBins - 1);	
+		// MOD-BY-LEETEN 04/11/2013:		uBin = min(max(uBin, 0), uNrOfBins - 1);	
+		uBin = min(max(uBin, (size_t)0), uNrOfBins - 1);	
+		// MOD-BY-LEETEN 04/11/2013-END
 		vfHist[uBin] += 1.0f;
 	}
